# public/docs/Developer/v1/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "emsearch API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https:\/\/emsearch.ryan.ems-dev.net"
        }
    ],
    "paths": {
...
        "\/api\/user\/{userId}": {
            "get": {
                "summary": "Get specified user",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "Update a specified user",
                "operationId": "putUpdate",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "user_group_id",
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "user_group_id": {
                                        "enum": [
                                            "Developer",
                                            "Support",
                                            "End-User"
                                        ],
                                        "type": "string"
                                    },
                                    "name": {
                                        "maxLength": 100,
                                        "type": "string"
                                    },
                                    "email": {
                                        "maxLength": 150,
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "Delete specified user",
                "description": "All relationships between the user and his projects will be automatically deleted too.<br \/>\nAll projects owned by the user will be automatically deleted too.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "Update a specified user",
                "operationId": "patchUpdate",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "user_group_id",
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "user_group_id": {
                                        "enum": [
                                            "Developer",
                                            "Support",
                                            "End-User"
                                        ],
                                        "type": "string"
                                    },
                                    "name": {
                                        "maxLength": 100,
                                        "type": "string"
                                    },
                                    "email": {
                                        "maxLength": 150,
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "\/api\/syncItem\/{syncItemId},{projectId}": {
            "get": {
                "summary": "Get specified sync item",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "syncItemId",
                        "in": "path",
                        "description": "Sync Item ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "$ref": "#\/components\/examples\/SyncItemShowExampleResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a specified sync item",
                "operationId": "putUpdate",
                "parameters": [
                    {
                        "name": "syncItemId",
                        "in": "path",
                        "description": "Sync Item ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "item_id",
                                    "project_id"
                                ],
                                "properties": {
                                    "item_id": {
                                        "maxLength": 191,
                                        "type": "string"
                                    },
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "item_signature": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "Delete specified sync item",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "syncItemId",
                        "in": "path",
                        "description": "Sync Item ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "Update a specified sync item",
                "operationId": "patchUpdate",
                "parameters": [
                    {
                        "name": "syncItemId",
                        "in": "path",
                        "description": "Sync Item ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "item_id",
                                    "project_id"
                                ],
                                "properties": {
                                    "item_id": {
                                        "maxLength": 191,
                                        "type": "string"
                                    },
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "item_signature": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "\/api\/syncTask\/{syncTaskId}": {
            "get": {
                "summary": "Get specified sync task",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "syncTaskId",
                        "in": "path",
                        "description": "Sync Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "Update a specified sync task",
                "operationId": "putUpdate",
                "parameters": [
                    {
                        "name": "syncTaskId",
                        "in": "path",
                        "description": "Sync Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "sync_task_type_id",
                                    "sync_task_status_id",
                                    "created_by_user_id",
                                    "project_id"
                                ],
                                "properties": {
                                    "sync_task_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "sync_task_type_id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    },
                                    "sync_task_status_id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    },
                                    "created_by_user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "planned_at": {
                                        "type": "string",
                                        "description": "Must be a valid date according to the strtotime PHP function."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "Delete specified sync task",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "syncTaskId",
                        "in": "path",
                        "description": "Sync Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "Update a specified sync task",
                "operationId": "patchUpdate",
                "parameters": [
                    {
                        "name": "syncTaskId",
                        "in": "path",
                        "description": "Sync Task ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "sync_task_type_id",
                                    "sync_task_status_id",
                                    "created_by_user_id",
                                    "project_id"
                                ],
                                "properties": {
                                    "sync_task_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "sync_task_type_id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    },
                                    "sync_task_status_id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    },
                                    "created_by_user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "planned_at": {
                                        "type": "string",
                                        "description": "Must be a valid date according to the strtotime PHP function."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "\/api\/syncTaskType\/{syncTaskTypeId}": {
            "get": {
                "summary": "Get specified sync task type",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "syncTaskTypeId",
                        "in": "path",
                        "description": "Sync task type ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "Update a sync task type",
                "operationId": "putUpdate",
                "parameters": [
                    {
                        "name": "syncTaskTypeId",
                        "in": "path",
                        "description": "Sync task type ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "Delete specified sync task type",
                "description": "The sync task type versions will be automatically deleted too.<br \/>\n<aside class=\"warning\">Avoid using this feature ! Check foreign keys constraints to remove dependent resources properly before.<\/aside>",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "syncTaskTypeId",
                        "in": "path",
                        "description": "Sync task type ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "Update a sync task type",
                "operationId": "patchUpdate",
                "parameters": [
                    {
                        "name": "syncTaskTypeId",
                        "in": "path",
                        "description": "Sync task type ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "\/api\/syncTaskStatus\/{syncTaskStatusId}": {
            "get": {
                "summary": "Get specified sync task status",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "syncTaskStatusId",
                        "in": "path",
                        "description": "Sync task type ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "Update a sync task status",
                "operationId": "putUpdate",
                "parameters": [
                    {
                        "name": "syncTaskStatusId",
                        "in": "path",
                        "description": "Sync task status ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "Delete specified sync task status",
                "description": "The sync task status versions will be automatically deleted too.<br \/>\n<aside class=\"warning\">Avoid using this feature ! Check foreign keys constraints to remove dependent resources properly before.<\/aside>",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "syncTaskStatusId",
                        "in": "path",
                        "description": "Sync task status ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "summary": "Update a sync task status",
                "operationId": "patchUpdate",
                "parameters": [
                    {
                        "name": "syncTaskStatusId",
                        "in": "path",
                        "description": "Sync task status ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "maxLength": 50,
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "apiOAuth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "\/oauth\/authorize",
                        "scopes": {
                            "*": "Access to anything"
                        }
                    },
                    "authorizationCode": {
                        "authorizationUrl": "\/oauth\/authorize",
                        "tokenUrl": "\/oauth\/token",
                        "refreshUrl": "\/oauth\/token\/refresh",
                        "scopes": {
                            "*": "Access to anything"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "apiBearer": [
                "*"
            ]
        },
        {
            "apiOAuth": [
                "*"
            ]
        }
    ]
}